refactor: pymc utils
